% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_doc.R
\name{include_doc}
\alias{include_doc}
\title{This function converts docx files to pdf and creates a labelled chunk title for them...the labelling doesn't work so is being manually set at the moment}
\usage{
include_doc(my_file, title = paste0(my_file), is_embed = TRUE)
}
\description{
This function converts docx files to pdf and creates a labelled chunk title for them...the labelling doesn't work so is being manually set at the moment
}
