% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_frag.R
\name{include_frag}
\alias{include_frag}
\title{Take docx file convert to markdown and include as child}
\usage{
include_frag(my_file)
}
\description{
Take docx file convert to markdown and include as child
}
